R-NaD/requirements.txt
absl-py==2.4.0
aiofiles==25.1.0
annotated-doc==0.0.4
anyio==4.13.0
attrs==26.1.0
certifi==2026.2.25
chex==0.1.85
click==8.3.1
cloudpickle==3.1.2
decorator==5.2.1
distrax==0.1.5
dm-env==1.6
dm-haiku==0.0.11
dm-tree==0.1.10
etils==1.14.0
filelock==3.25.2
flax==0.12.0
fsspec==2026.2.0
gast==0.7.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
humanize==4.15.0
idna==3.11
jax==0.9.2
jax-cuda12-pjrt==0.9.2
jax-cuda12-plugin==0.9.2
jaxlib==0.9.2
Jinja2==3.1.6
jmp==0.0.4
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
msgpack==1.1.2
networkx==3.6.1
numpy==2.4.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.20.0.48
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opt_einsum==3.4.0
optax==0.1.9
orbax-checkpoint==0.11.33
packaging==26.0
pillow==12.1.1
protobuf==7.34.1
psutil==7.2.2
Pygments==2.20.0
PyYAML==6.0.3
regex==2026.3.32
rich==14.3.3
rlax==0.1.7
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.3.0
setuptools==70.2.0
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
sympy==1.13.1
tabulate==0.10.0
tensorflow-probability==0.25.0
tensorstore==0.1.82
threadpoolctl==3.6.0
tokenizers==0.22.2
toolz==1.1.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.3
transformers==5.4.0
treescope==0.1.10
triton==3.1.0
typer==0.24.1
typing_extensions==4.15.0
uvloop==0.22.1
wrapt==2.1.2
zipp==3.23.0
